requirements.txt
telebot==0.0.5
Flask==3.0.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6